Refactorisation des noms et emplacements des fichiers
