+++ requirements.txt
@@ -1,0 +1,5 @@
+dash
+plotly
+pandas
+dash_mantine_components
+pycountry
